+++ components/cart/CartList/CartItem.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { CartType } from '@type/cartType';
+
+interface CartItemProps {
+  cart: CartType;
+}
+
+export default function CartItem({ cart }: CartItemProps) {
+  return <div>CartItem</div>;
+}
