+++ components/cart/CartList/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function CartList() {
+  return <div>CartList</div>;
+}
