+++ components/shop/ProductDetail/ProductCartItem.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { MOCK_PRODUCT_CART } from '@constants/mockData';
+import { MOCK_PRODUCT_CART } from '@constants/mock/productCart';
 import ProductCartItem from './ProductCartItem';
 
 const meta: Meta<typeof ProductCartItem> = { component: ProductCartItem };
