+++ components/shop/ProductDetail/ProductCartList.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { MOCK_PRODUCT_CART } from '@constants/mockData';
+import { MOCK_PRODUCT_CART } from '@constants/mock/productCart';
 import ProductCartList from './ProductCartList';
 
 const meta: Meta<typeof ProductCartList> = { component: ProductCartList };
