+++ constants/mock/productList.ts
@@ -1,0 +1,155 @@
+import { createRandomProductTitle, generateRandomId } from '@utils/mock';
+import { ProductItemPreview } from '@type/productType';
+
+export const MOCK_PREVIEW_PRODUCT_LIST: ProductItemPreview[] = [
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+    category: 'best',
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+    category: 'new',
+    discountPercent: Math.floor(Math.random() * 100),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+    category: 'best',
+    discountPercent: Math.floor(Math.random() * 100),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+    discountPercent: Math.floor(Math.random() * 100),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+    discountPercent: Math.floor(Math.random() * 100),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+  {
+    id: generateRandomId(16),
+    hoverImageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    imageUrl: `https://source.unsplash.com/600x400/?nature,water&sig=${Math.floor(
+      Math.random() * 100
+    )}`,
+    price: Math.floor(Math.random() * 10000),
+    title: createRandomProductTitle(),
+  },
+];
